+++ priv/templates/brando.gen/admin/list.ex
@@ -1,5 +1,5 @@
 defmodule <%= app_module %>Admin.<%= domain %>.<%= camel_singular %>ListLive do
-  use BrandoAdmin.LiveView.Listing, schema: <%= inspect schema_module %>
+  use BrandoAdmin.LiveView.Listing.Compiler, schema: <%= inspect schema_module %>
   use Gettext, backend: <%= admin_module %>.Gettext, warn: false
   alias BrandoAdmin.Components.Content